# docs/swagger.yaml
openapi: 3.0.0
paths:
...
  /v2/queue/callback/{task}:
    post:
      operationId: QueueController_handleCallback
      parameters:
        - 
          name: task
          required: true
          in: path
          schema:
            type: string
      responses:
        201:
          description: ""
  /v2/members/committee:
    get:
      operationId: MembersController_getCommitteeMembers
      summary: Get committee/council members (public)
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommitteeMemberDto"
      tags:
        - Members
      security:
        - 
          bearer: []
  /v2/members/notable:
    get:
      operationId: MembersController_getNotableMembers
      summary: Get notable/distinguished members (public)
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotableMemberDto"
      tags:
        - Members
      security:
        - 
          bearer: []
  /v2/members/stats:
    get:
      operationId: MembersController_getMemberStats
      summary: Aggregate member stats for header tiles (public)
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberStatsResponseDto"
      tags:
        - Members
      security:
        - 
          bearer: []
  /v2/members:
    get:
      operationId: MembersController_getAllMembers
      summary: Get all members (paginated, optional name filter for the BO table)
      parameters:
        - 
          name: page
          required: false
          in: query
          schema:
            type: number
        - 
          name: pageSize
          required: false
          in: query
          schema:
            type: number
        - 
          name: name
          required: false
          in: query
          description: Filter by name (Vietnamese-insensitive), used by the members table search
          schema:
            type: string
        - 
          name: sortBy
          required: false
          in: query
          schema:
            enum:
              - created_at
              - name
              - generation
            type: string
        - 
          name: sortOrder
          required: false
          in: query
          schema:
            enum:
              - asc
              - desc
            type: string
        - 
          name: view
          required: false
          in: query
          description: "Độ chi tiết của payload. Bỏ trống = `full` (mặc định, KHÔNG đổi shape).\n- `full`  → PaginatedMembersResponseDto (mặc định)\n- `table` → PaginatedMembersTableResponseDto (bỏ biography/notes — dùng cho bảng data BO)\n- `lite`  → PaginatedMembersLiteResponseDto (id + name + avatar + generation)"
          schema:
            enum:
              - lite
              - table
              - full
            type: string
        - 
          name: tree_id
          required: false
          in: query
          description: Lọc theo chi nhánh (UUID, FK tới Tree). Dropdown lấy từ GET /v2/tree.
          schema:
            type: string
        - 
          name: gender
          required: false
          in: query
          description: Lọc theo giới tính
          schema:
            enum:
              - M
              - F
              - O
              - U
            type: string
        - 
          name: generation
          required: false
          in: query
          description: Lọc theo thế hệ (giá trị hiệu lực trên member)
          schema:
            type: number
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMembersResponseDto"
      tags:
        - Members
      security:
        - 
          bearer: []
    post:
      operationId: MembersController_createMember
      summary: Create new member + profile (editor trở lên)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponseDto"
      tags:
        - Members
      security:
        - 
          bearer: []
  /v2/members/generations/recompute:
    post:
      operationId: MembersController_recomputeGenerations
      summary: Tính lại thế hệ cho toàn bộ thành viên (admin). Job nền vốn đã tự chạy sau mỗi lần ghi.
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecomputeGenerationsResponseDto"
      tags:
        - Members
      security:
        - 
          bearer: []
  /v2/members/search:
    get:
      operationId: MembersController_searchMembers
      summary: Search tên cho select/autocomplete (unpaginated, trả MemberLiteDto)
      parameters:
        - 
          name: name
          required: true
          in: query
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberLiteDto"
      tags:
        - Members
      security:
        - 
          bearer: []
  /v2/members/{id}:
    get:
      operationId: MembersController_getMemberById
      summary: Get member basic info
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponseDto"
      tags:
        - Members
      security:
        - 
          bearer: []
    delete:
      operationId: MembersController_deleteMember
      summary: "Delete member (cascade: profile, userMetadata)"
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Members
      security:
        - 
          bearer: []
  /v2/members/{id}/profile:
    get:
      operationId: MembersController_getMemberProfile
      summary: Get member with full profile + relationships
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfileResponseDto"
      tags:
        - Members
      security:
        - 
          bearer: []
    put:
      operationId: MembersController_updateMemberProfile
      summary: "Update member profile (member: chỉ hồ sơ của chính mình + field cho phép; editor trở lên: mọi người, mọi field)"
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateMemberDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponseDto"
      tags:
        - Members
      security:
        - 
          bearer: []
...
  /v2/media/member/{memberId}:
    get:
      operationId: MediaController_getMemberMedia
      summary: Tài liệu/ảnh gắn với một thành viên (Tài liệu liên quan)
      parameters:
        - 
          name: memberId
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaResponseDto"
      tags:
        - Media
      security:
        - 
          bearer: []
...
  /v2/members/{memberId}/memories:
    get:
      operationId: MemoriesController_getByMember
      summary: Get memories attached to a member (public)
      parameters:
        - 
          name: memberId
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryResponseDto"
      tags:
        - Memories (Kỷ niệm)
      security:
        - 
          bearer: []
    post:
      operationId: MemoriesController_create
      summary: Add a memory to a member (author = current user)
      parameters:
        - 
          name: memberId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryResponseDto"
      tags:
        - Memories (Kỷ niệm)
      security:
        - 
          bearer: []
  /v2/members/{memberId}/memories/{id}:
    delete:
      operationId: MemoriesController_delete
      summary: Delete a memory
      parameters:
        - 
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Deleted
      tags:
        - Memories (Kỷ niệm)
      security:
        - 
          bearer: []
...
info:
  title: Family Tree API v2
  description: "## Vietnamese Family Tree Management API\n\n**Base URL:** `http://localhost:3002/v2`\n\n**Authentication:** Bearer JWT token (same as backend v1)\n\n### Modules\n- **Members** — CRUD thành viên + profile + avatar (async upload)\n- **Relationships** — Quan hệ mới (BIOLOGICAL/ADOPTED/SPOUSE) + tìm tổ tiên/con cháu\n- **Tree** — Cây gia phả full (Redis cache 1h) + subtree 4 thế hệ\n- **Anniversaries** — Ngày giỗ (filter by member, month, upcoming)\n- **Events** — Sự kiện dòng họ + notification queue\n- **Media** — Thư viện media: upload mọi loại (ảnh nén lossless bằng sharp; video/audio/tài liệu upload thẳng) → Vercel Blob; phân trang/lọc/tìm kiếm + thống kê + album\n- **Graves** — Mộ phần với tọa độ GPS + tìm kiếm gần nhất\n\n### Queue Jobs (QStash + Redis)\n| Queue | Trigger | Action |\n|-------|---------|--------|\n| avatar-upload | Create/Update member với file | Upload → Vercel Blob → cập nhật avatar_url |\n| media-process | Upload media | Ảnh: sharp nén lossless; khác: upload thẳng → Vercel Blob (off-request) |\n| report-generate | Create/Delete member | Tính stats → lưu Redis |\n| notification | New member/relationship/event | Log (Phase 1) |\n"
  version: 2.0.0
  contact:
    name: Family Tree Team
    url: ""
    email: ""
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      in: header
      description: Dán CHỈ token thô (lấy từ POST /v2/auth/login), KHÔNG kèm tiền tố "Bearer ".
  schemas:
...
    MemberLiteDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Nguyễn Văn A
        avatar_url:
          type: object
          nullable: true
          format: uri
        generation:
          type: number
          nullable: true
          example: 5
          description: Thế hệ hiệu lực (đã tính sẵn). Xem MemberResponseDto.generation.
      required:
        - id
        - name
...
    CommitteeMemberDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          example: Trưởng ban
          description: Lấy từ profile.occupation
        avatar:
          type: string
          format: uri
      required:
        - id
        - name
        - role
        - avatar
    NotableMemberDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          description: Lấy từ profile.biography
        avatar:
          type: string
          format: uri
      required:
        - id
        - name
        - description
        - avatar
    MemberStatsResponseDto:
      type: object
      properties:
        total:
          type: number
          example: 1256
        male:
          type: number
          example: 648
        female:
          type: number
          example: 608
        newThisMonth:
          type: number
          example: 28
          description: Thành viên tạo mới trong tháng hiện tại
        generations:
          type: number
          example: 6
          description: Số thế hệ (distinct generation)
      required:
        - total
        - male
        - female
        - newThisMonth
        - generations
...
    MemberResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        private_id:
          type: object
          nullable: true
        name:
          type: string
          example: Nguyễn Văn A
        normalized_name:
          type: object
          nullable: true
          description: Tên đã bỏ dấu để tìm kiếm
        avatar_url:
          type: object
          nullable: true
          format: uri
        gender:
          type: string
          nullable: true
          enum:
            - M
            - F
            - O
            - U
        birthDate:
          type: object
          nullable: true
          example: 1990-01-01
        deathDate:
          type: object
          nullable: true
          example: 2020-12-31
        tree_id:
          type: object
          nullable: true
          format: uuid
          description: Chi nhánh — FK tới Tree
        generation:
          type: number
          nullable: true
          example: 5
          description: "Thế hệ (giá trị HIỆU LỰC, đã tính sẵn) = `profile.generation` nếu nhập tay, ngược lại suy ra từ quan hệ cha-con/vợ-chồng. Cập nhật bởi job nền `generation-recompute`. Muốn biết là nhập tay hay suy ra: `profile.generation != null`."
        generation_updated_at:
          type: string
          nullable: true
          format: date-time
          description: Lần cuối job nền ghi lại `generation`
        created_at:
          type: string
          format: date-time
        profile:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/ProfileResponseDto"
        tree:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/MemberTreeBriefDto"
      required:
        - id
        - name
        - created_at
    PaginatedMembersResponseDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MemberResponseDto"
        total:
          type: number
          example: 120
          description: Tổng số thành viên
        page:
          type: number
          example: 1
        pageSize:
          type: number
          example: 10
      required:
        - data
        - total
        - page
        - pageSize
    RecomputeGenerationsResponseDto:
      type: object
      properties:
        members:
          type: number
          example: 1256
          description: Số thành viên đã duyệt
        updated:
          type: number
          example: 12
          description: Số dòng thực sự đổi giá trị
        durationMs:
          type: number
          example: 840
        warnings:
          description: Bất thường trong dữ liệu (cycle, pin mâu thuẫn, cạnh tự trỏ)
          type: array
          items:
            type: string
      required:
        - members
        - updated
        - durationMs
        - warnings
    CreateMemberDto:
      type: object
      properties:
        fullName:
          type: string
          example: Nguyễn Văn A
        gender:
          type: string
          example: M
          enum:
            - M
            - F
            - O
            - U
        birthDate:
          type: string
          example: 1990-01-01
        deathDate:
          type: string
          example: 2020-12-31
        occupation:
          type: string
          example: Kỹ sư phần mềm
        address:
          type: string
          example: Hà Nội
        biography:
          type: string
        generation:
          type: number
        clanRole:
          type: string
          enum:
            - TRUONG_TOC
            - PHO_TRUONG_TOC
            - THANH_VIEN
          description: Vai trò trong tộc (map vào committeeRole/isCommittee)
        phone:
          type: string
          example: 0988 123 456
        contactEmail:
          type: string
          example: lienhe@gmail.com
          description: Email liên hệ (khác email tài khoản)
        tree_id:
          type: string
          format: uuid
          description: Chi nhánh — FK tới Tree
        familyPosition:
          type: string
          example: Con trưởng
          description: Vị trí trong gia đình
        roleTags:
          description: Vai trò trong dòng họ (badges)
          type: array
          items:
            type: string
      required:
        - fullName
        - gender
...
    MemberProfileResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        private_id:
          type: object
          nullable: true
        name:
          type: string
          example: Nguyễn Văn A
        normalized_name:
          type: object
          nullable: true
          description: Tên đã bỏ dấu để tìm kiếm
        avatar_url:
          type: object
          nullable: true
          format: uri
        gender:
          type: string
          nullable: true
          enum:
            - M
            - F
            - O
            - U
        birthDate:
          type: object
          nullable: true
          example: 1990-01-01
        deathDate:
          type: object
          nullable: true
          example: 2020-12-31
        tree_id:
          type: object
          nullable: true
          format: uuid
          description: Chi nhánh — FK tới Tree
        generation:
          type: number
          nullable: true
          example: 5
          description: "Thế hệ (giá trị HIỆU LỰC, đã tính sẵn) = `profile.generation` nếu nhập tay, ngược lại suy ra từ quan hệ cha-con/vợ-chồng. Cập nhật bởi job nền `generation-recompute`. Muốn biết là nhập tay hay suy ra: `profile.generation != null`."
        generation_updated_at:
          type: string
          nullable: true
          format: date-time
          description: Lần cuối job nền ghi lại `generation`
        created_at:
          type: string
          format: date-time
        profile:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/ProfileResponseDto"
        tree:
          nullable: true
          allOf:
            - 
              $ref: "#/components/schemas/MemberTreeBriefDto"
        parent_relationships:
          description: Quan hệ mà member là con (chứa parent)
          type: array
          items:
            $ref: "#/components/schemas/ProfileRelationshipDto"
        child_relationships:
          description: Quan hệ mà member là cha/mẹ (chứa child)
          type: array
          items:
            $ref: "#/components/schemas/ProfileRelationshipDto"
      required:
        - id
        - name
        - created_at
        - parent_relationships
        - child_relationships
    UpdateMemberDto:
      type: object
      properties:
        fullName:
          type: string
        gender:
          type: string
          enum:
            - M
            - F
            - O
            - U
        birthDate:
          type: string
        deathDate:
          type: string
        occupation:
          type: string
        address:
          type: string
        biography:
          type: string
        notes:
          type: string
        generation:
          type: number
        clanRole:
          type: string
          enum:
            - TRUONG_TOC
            - PHO_TRUONG_TOC
            - THANH_VIEN
          description: Vai trò trong tộc (map vào committeeRole/isCommittee)
        phone:
          type: string
          example: 0988 123 456
        contactEmail:
          type: string
          example: lienhe@gmail.com
        tree_id:
          type: string
          format: uuid
          description: Chi nhánh — FK tới Tree
        familyPosition:
          type: string
          example: Con trưởng
        roleTags:
          type: array
          items:
            type: string
...
    MediaResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        file_path:
          type: string
          format: uri
          description: URL Blob khi ready; /pending/... khi đang xử lý
        title:
          type: object
          nullable: true
          example: Lễ giỗ Tổ 2024
        type:
          type: string
          nullable: true
          enum:
            - image
            - video
            - audio
            - document
        mime_type:
          type: object
          nullable: true
          example: image/jpeg
        size_bytes:
          type: number
          nullable: true
          example: 2411724
        duration_seconds:
          type: number
          nullable: true
          example: 324
          description: Thời lượng (giây)
        views:
          type: number
          example: 2456
          description: Lượt xem
        tags:
          default: []
          type: array
          items:
            type: string
        status:
          type: string
          enum:
            - pending
            - ready
            - failed
          example: ready
        member_id:
          type: object
          nullable: true
          format: uuid
        album_id:
          type: object
          nullable: true
          format: uuid
        event_id:
          type: object
          nullable: true
          format: uuid
        uploader_id:
          type: string
          format: uuid
        uploader_name:
          type: object
          nullable: true
          example: Nguyễn Văn An
        created_at:
          type: string
          format: date-time
      required:
        - id
        - file_path
        - views
        - tags
        - status
        - uploader_id
        - created_at
...
    MemoryResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        member_id:
          type: object
          nullable: true
          format: uuid
        event_id:
          type: object
          nullable: true
          format: uuid
        author_id:
          type: string
          format: uuid
        text:
          type: string
        photos:
          default: []
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - author_id
        - text
        - photos
        - created_at
    CreateMemoryDto:
      type: object
      properties:
        text:
          type: string
          example: Kỷ niệm khó quên trong ngày họp mặt dòng họ...
        photos:
          description: Ảnh kỷ niệm
          type: array
          items:
            type: string
            format: uri
        event_id:
          type: string
          format: uuid
          description: Gắn với một sự kiện (tùy chọn)
      required:
        - text
...